refactor(sql): lift window primitives to a shared sql_window module
CurrentRow, Unbounded, Preceding, Following, OrderBy, WindowFrame, and the
_render_over_clause / _render_frame_bound / _render_order_by_item / _bound_rank
helpers move from duckdb_relation.py into a new sql/sql_window.py sibling.
They are pure SQL-fragment generators with no DuckDB-specific logic, so
lifting them lets SqliteRelation reuse the same primitives. duckdb_relation
re-exports the public types to preserve its existing __all__ and import
surface for callers.
